+++ packages/terra-action-header/tests/jest/ActionHeader.test.jsx
@@ -10,43 +10,43 @@ describe('ActionHeader', () => {
   it('should render a default action header', () => {
     const actionHeader = <ActionHeader />;
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an action header with title', () => {
     const actionHeader = <ActionHeader title="Action Header" />;
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an action header with back button and title', () => {
     const actionHeader = <ActionHeader title="Action Header" onBack={() => {}} />;
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an action header with close button and title', () => {
     const actionHeader = <ActionHeader title="Action Header" onClose={() => {}} />;
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an action header with back and close buttons and title', () => {
     const actionHeader = <ActionHeader title="Action Header" onBack={() => {}} onClose={() => {}} />;
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an action header with level three header element and title', () => {
     const actionHeader = <ActionHeader title="Action Header" level={3} />;
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an action header with custom button and title', () => {
     const actionHeader = <ActionHeader title="Action Header"><Button text="Custom Button" onClick={() => alert('You clicked me!')} /></ActionHeader>;
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an action header with multiple custom buttons and title', () => {
@@ -59,36 +59,36 @@ describe('ActionHeader', () => {
       </ActionHeader>
     );
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an action header with maximize button and title', () => {
     const actionHeader = <ActionHeader title="Action Header" onMaximize={() => {}} />;
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an action header with minimize button and title', () => {
     const actionHeader = <ActionHeader title="Action Header" onMinimize={() => {}} />;
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an action header with next and previous buttons and title', () => {
     const actionHeader = <ActionHeader title="Action Header" onNext={() => {}} onPrevious={() => {}} />;
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an action header with title, enabled next button, and disabled previous button', () => {
     const actionHeader = <ActionHeader title="Action Header" onNext={() => {}} />;
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an action header with title, enabled previous button, and disabled next button', () => {
     const actionHeader = <ActionHeader title="Action Header" onPrevious={() => {}} />;
     const wrapper = shallowWithIntl(actionHeader);
-    expect(wrapper.dive()).toMatchSnapshot();
+    expect(wrapper).toMatchSnapshot();
   });
 });