+++ packages/terra-overlay/tests/jest/LoadingOverlay.test.jsx
@@ -5,7 +5,7 @@ import LoadingOverlay from '../../src/LoadingOverlay';
 
 describe('LoadingOverlay', () => {
   it('should not render markup when isOpen is not provided', () => {
-    const wrapper = mountWithIntl(<LoadingOverlay />).children();
+    const wrapper = mountWithIntl(<LoadingOverlay />);
     expect(wrapper).toMatchSnapshot();
   });
 
@@ -14,19 +14,19 @@ describe('LoadingOverlay', () => {
       const defaultRender = <LoadingOverlay isOpen />;
 
       it('should render a default component', () => {
-        const wrapper = mountWithIntl(defaultRender).children();
+        const wrapper = mountWithIntl(defaultRender);
         expect(wrapper).toMatchSnapshot();
       });
 
       it('should have the class loading-overlay', () => {
-        const wrapper = mountWithIntl(defaultRender).children();
+        const wrapper = mountWithIntl(defaultRender);
         expect(wrapper.find('loading-overlay'));
       });
     });
 
     it('should render with message prop', () => {
       const overlayRender = <LoadingOverlay isOpen message="Loading!" />;
-      const wrapper = mountWithIntl(overlayRender).children();
+      const wrapper = mountWithIntl(overlayRender);
       expect(wrapper.find('.message').text()).toEqual('Loading!');
       expect(wrapper).toMatchSnapshot();
     });