+++ packages/terra-form-radio/CHANGELOG.md
@@ -3,7 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
-
+### Changed
+* Components updated to use `FormattedMessage` to interface with `react-intl's` `intl` context.
 3.23.0 - (July 30, 2019)
 ------------------
 ### Changed