+++ packages/terra-form-radio/CHANGELOG.md
@@ -5,6 +5,8 @@ Unreleased
 ----------
 ### Changed
 * Components updated to use `FormattedMessage` to interface with `react-intl's` `intl` context.
+* updated package.json test scripts
+
 3.23.0 - (July 30, 2019)
 ------------------
 ### Changed